+++ src/components/Carousel/Carousel.less
@@ -1,0 +1,5 @@
+.carousel {
+    width: 1024px;
+    height: 450px;
+    margin: 0 auto;
+}
