+++ src/less/slide.less
@@ -2,7 +2,6 @@
     background: url('../../public/assets/img/bg.png') repeat;
 
     .slide {
-        height: 450px;
 
         .slide-img {
             float: left;
